Rewrite timer code to have an always-running timer and maintain our own count.
The previous timer code relied on resetting alarm times within the timer
interrupt. It turns out, the ESP32 code (whether IDF or Arduino) doesn't allow
this. We could apparently have achieved the same thing by directly manipulating
registers, but that doesn't sound like fun. Instead, we now have a timer
interrupt continually running and maintain our own counter to indicate whether
it should trigger action or not. This counter can be reset from within an
interrupt without problems. The timer is theoretically a bit wasteful but
a) any DALI adapter will be running with mains power anyway, so this isn't
   draining any batteries
b) in the "inactive" case, the interrupt is one conditional and a return