Explicitly mark whether a sent message expects a back frame or not.
Our previous condition was correct about the range of messages that expect a
back frame, but didn't deal with messages using ADDR_COMPARE or
ADDR_VERIFY_SHORT_ADDR (both of which expect a back frame, but you can't tell
that from looking at the msg) field, nor with messages such as
ADDR_SEARCH_ADDR_H which might fill the msg field in such a way that it looks
like they expect a back frame, but actually don't. Explicitly tracking this is
a lot easier than writing the enormous if condition that would otherwise be
needed.