Explicitly track send_twice.
As previously for back frames, the if condition for if something needs to be
sent twice is too much like hard work because sometimes a msg value will look
like it needs to be sent twice, but it's actually part of a search addr h
message and doesn't need to be. Track it explicitly instead.

Also, note all messages (whether by message type or addr) that need to be sent
twice and those that produce back frames.